# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: FETAL-GUARD API
  description: |
    Backend API for FETAL-GUARD: Wearable Early Warning System for fetal monitoring.
    Provides endpoints for patient management, device pairing, monitoring sessions,
    alerts, and clinician dashboard.
  version: 1.0.0
  contact:
    name: FETAL-GUARD Team
    email: support@fetal-guard.id
  license:
    name: Proprietary
servers:
  - url: https://api.fetal-guard.id/v1
    description: Production server
  - url: https://staging-api.fetal-guard.id/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: User login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
        user:
          type: object
          properties:
            id:
              type: string
            email:
              type: string
            role:
              type: string
              enum: [patient, clinician, admin]
...